Monitoring_Theft/requirements.txt
# python = 3.11.11
sqlalchemy
python-dotenv
psycopg2-binary
ultralytics
pandas
fastapi
uvicorn
python-multipart
httpx==0.28.1
scikit-learn==1.6.1
